requirements.txt
openpyxl
beautifulsoup4
lxml
virtualenv
pyyaml
lastversion
fosslight_util>=1.4.21
PyGithub
requirements-parser
